move runner-ct controller into server-ct and add proper types
